requirements.txt
discord.py==1.5.1
ruamel.yaml==0.16.12 